transferservice/requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.4.0
Flask==1.1.1
Flask-Cors==3.0.8
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-core==1.0.3
google-cloud-datastore==1.10.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
gunicorn==20.0.2
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
MarkupSafe==1.1.1
more-itertools==7.2.0
openapi-spec-validator==0.2.8
pathspec==0.6.0
protobuf==3.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyrsistent==0.15.6
pytz==2019.3
PyYAML==5.1.2
regex==2019.11.1
requests==2.22.0
rsa==4.0
six==1.13.0
swagger-ui-bundle==0.0.6
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.7
Werkzeug==0.16.0
zipp==0.6.0
